app: restructure files based on Buffer's package structure
