+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
     android:supportsRtl="true"
     android:theme="@style/Theme.MaterialComponents">
     <activity
-      android:name=".MainActivity"
+      android:name=".ui.main.MainActivity"
       android:label="@string/app_name"
       android:theme="@style/Theme.MaterialComponents.NoActionBar">
       <intent-filter>
