+++ app/src/main/java/dev/msfjarvis/lobsters/ui/urllauncher/UrlLauncherAmbient.kt
@@ -1,4 +1,4 @@
-package dev.msfjarvis.lobsters.urllauncher
+package dev.msfjarvis.lobsters.ui.urllauncher
 
 import androidx.compose.runtime.ambientOf
